[front] cleanup: colocate feedback distribution tooltip
